chore: tweak sauce configuration
